Fix mocking. Also use existing StorageStatement
