+++ community/kernel/src/test/java/org/neo4j/kernel/api/KernelTransactionFactory.java
@@ -37,6 +37,7 @@ import org.neo4j.kernel.impl.locking.NoOpClient;
 import org.neo4j.kernel.impl.locking.SimpleStatementLocks;
 import org.neo4j.kernel.impl.locking.StatementLocks;
 import org.neo4j.kernel.impl.proc.Procedures;
+import org.neo4j.kernel.impl.store.StoreHolder;
 import org.neo4j.kernel.impl.transaction.TransactionHeaderInformationFactory;
 import org.neo4j.kernel.impl.transaction.TransactionMonitor;
 import org.neo4j.resources.CpuClock;
@@ -80,10 +81,12 @@ public class KernelTransactionFactory
         when( headerInformationFactory.create() ).thenReturn( headerInformation );
 
         StorageEngine storageEngine = mock( StorageEngine.class );
+        StoreHolder storeHolder = mock( StoreHolder.class );
         StoreReadLayer storeReadLayer = mock( StoreReadLayer.class );
         StorageStatement storageStatement = mock( StorageStatement.class );
         when( storeReadLayer.newStatement() ).thenReturn( storageStatement );
         when( storageEngine.storeReadLayer() ).thenReturn( storeReadLayer );
+        when( storageEngine.stores() ).thenReturn( storeHolder );
 
         KernelTransactionImplementation transaction = new KernelTransactionImplementation(
                 mock( StatementOperationParts.class ),
