+++ enterprise/kernel/src/test/java/org/neo4j/kernel/enterprise/builtinprocs/TransactionStatusResultTest.java
@@ -20,23 +20,18 @@
 package org.neo4j.kernel.enterprise.builtinprocs;
 
 import org.junit.Test;
-import org.mockito.Answers;
 
 import java.net.InetSocketAddress;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Stream;
 
 import org.neo4j.internal.kernel.api.security.SecurityContext;
 import org.neo4j.io.pagecache.tracing.cursor.PageCursorTracer;
 import org.neo4j.kernel.api.KernelTransaction;
 import org.neo4j.kernel.api.KernelTransactionHandle;
-import org.neo4j.kernel.api.Statement;
 import org.neo4j.kernel.api.exceptions.InvalidArgumentsException;
-import org.neo4j.kernel.api.exceptions.Status;
-import org.neo4j.kernel.api.exceptions.TransactionFailureException;
 import org.neo4j.kernel.api.query.ExecutingQuery;
 import org.neo4j.kernel.api.query.QuerySnapshot;
 import org.neo4j.kernel.impl.api.KernelTransactionImplementation;
@@ -189,115 +184,6 @@ public class TransactionStatusResultTest
                     statistics );
             return new TransactionExecutionStatistic( transaction, Clocks.fakeClock().forward( 2010, MILLISECONDS ),
                     200 );
-        }
-    }
-
-    private static class TestKernelTransaction implements KernelTransaction
-    {
-        @Override
-        public Statement acquireStatement()
-        {
-            return null;
-        }
-
-        @Override
-        public void success()
-        {
-        }
-
-        @Override
-        public void failure()
-        {
-        }
-
-        @Override
-        public long closeTransaction() throws TransactionFailureException
-        {
-            return 0;
-        }
-
-        @Override
-        public boolean isOpen()
-        {
-            return false;
-        }
-
-        @Override
-        public SecurityContext securityContext()
-        {
-            SecurityContext securityContext = mock( SecurityContext.class, Answers.RETURNS_DEEP_STUBS );
-            when( securityContext.subject().username() ).thenReturn( "testUser" );
-            return securityContext;
-        }
-
-        @Override
-        public Optional<Status> getReasonIfTerminated()
-        {
-            return Optional.empty();
-        }
-
-        @Override
-        public boolean isTerminated()
-        {
-            return false;
-        }
-
-        @Override
-        public void markForTermination( Status reason )
-        {
-        }
-
-        @Override
-        public long lastTransactionTimestampWhenStarted()
-        {
-            return 0;
-        }
-
-        @Override
-        public long lastTransactionIdWhenStarted()
-        {
-            return 0;
-        }
-
-        @Override
-        public long startTime()
-        {
-            return 1984;
-        }
-
-        @Override
-        public long timeout()
-        {
-            return 0;
-        }
-
-        @Override
-        public void registerCloseListener( CloseListener listener )
-        {
-        }
-
-        @Override
-        public Type transactionType()
-        {
-            return null;
-        }
-
-        @Override
-        public long getTransactionId()
-        {
-            return 8;
-        }
-
-        @Override
-        public long getCommitTime()
-        {
-            return 0;
-        }
-
-        @Override
-        public Revertable overrideWith( SecurityContext context )
-        {
-            return null;
         }
     }
